refactor(listing purpose): renamed listingPurposeKey to listingPurpose on subgroup schema
