+++ packages/cp-listing-service/src/routes/settings/subgroup.routes.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { getAllSubgroupsHandler } from '../../controllers/settings/purpose.controllers';
+
+const router = Router();
+
+router.get('/', getAllSubgroupsHandler);
+
+export { router as subgroupRoutes };
